HUDManager refactor
Nauja HUDManager klasė, valdo HUD langų būsenas, nebereikia padavinėti reference į skirtingus VM
