+++ src/TinyCiv.Client/Code/MVVM/MainViewModel.cs
@@ -19,20 +19,15 @@ namespace TinyCiv.Client.Code.MVVM
     public class MainViewModel
     {
         public GameViewModel GameVM = new GameViewModel();
-        public UpperMenuViewModel UpperMenuVM = new UpperMenuViewModel();
-        public UnitMenuViewModel UnitMenuVM = new UnitMenuViewModel();
 
         public ObservableValue<object> Game { get; } = new ObservableValue<object>();
         public ObservableValue<object> UpperMenu { get; } = new ObservableValue<object>();
         public ObservableValue<object> LowerMenu { get; } = new ObservableValue<object>(new LobbyMenuViewModel());
 
-        public ObservableValue<String> IsUnitStatVisible { get; } = new ObservableValue<string>("Hidden");
-
-
         public MainViewModel()
         {
             Game.Value = GameVM;
-            //UpperMenu.Value = UpperMenuVM;
+            HUDManager.mainVM = this;
 
             DependencyObject dep = new DependencyObject();
             if (!DesignerProperties.GetIsInDesignMode(dep))
@@ -53,7 +48,7 @@ namespace TinyCiv.Client.Code.MVVM
             if (CurrentPlayer.Instance.player == null)
             {
                 CurrentPlayer.Instance.player = response.Created;
-                UpperMenuVM.PlayerColor.Value = CurrentPlayer.Color;
+                HUDManager.DisplayUpperMenu();
             }
 
             // If the party is full
@@ -64,13 +59,8 @@ namespace TinyCiv.Client.Code.MVVM
 
         private void OnGameStart(GameStartServerEvent response)
         {
-            LowerMenu.Value = UnitMenuVM;
-            UpperMenu.Value = UpperMenuVM;
-
             GameVM.GameStart(response);
-
-            GameVM.UnitMenuVM = UnitMenuVM;
-            GameVM.UpperMenuVM = UpperMenuVM;
+            HUDManager.DisplayCityMenu();
         }
 
     }
