+++ src/TinyCiv.Client/Code/MVVM/ViewModel/UpperMenuViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using TinyCiv.Shared;
+using TinyCiv.Shared.Events.Server;
 using TinyCiv.Shared.Game;
 
 namespace TinyCiv.Client.Code.MVVM.ViewModel
@@ -10,8 +11,15 @@ namespace TinyCiv.Client.Code.MVVM.ViewModel
         public ObservableValue<int> Food { get; } = new ObservableValue<int>(Constants.Game.StartingFood);
         public ObservableValue<TeamColor> PlayerColor { get; } = new ObservableValue<TeamColor>();
 
-        public void SetResources(Resources resources)
+        public UpperMenuViewModel()
         {
+            PlayerColor.Value = CurrentPlayer.Color;
+            ClientSingleton.Instance.serverClient.ListenForResourcesUpdate(OnResourceUpdate);
+        }
+
+        public void OnResourceUpdate(ResourcesUpdateServerEvent response)
+        {
+            var resources = response.Resources;
             Gold.Value = resources.Gold;
             Industry.Value = resources.Industry;
             Food.Value = resources.Food;
